Refactoring: moved provider extender into separate folder
