+++ internal/controller/runtime/fsm/runtime_fsm_patch_shoot.go
@@ -48,7 +48,7 @@ func sFnPatchExistingShoot(ctx context.Context, m *fsm, s *systemState) (stateFn
 		Resources:            s.shoot.Spec.Resources,
 		InfrastructureConfig: s.shoot.Spec.Provider.InfrastructureConfig,
 		ControlPlaneConfig:   s.shoot.Spec.Provider.ControlPlaneConfig,
-		Log: ptr.To(m.log),
+		Log:                  ptr.To(m.log),
 	})
 
 	if err != nil {
