+++ pkg/gardener/shoot/extender/provider/provider_gcp_test.go
@@ -1,4 +1,4 @@
-package extender
+package provider
 
 import (
 	"encoding/json"
